+++ examples/intro.py
@@ -1,43 +1,22 @@
-# pylint: disable=missing-return-doc, missing-param-doc, missing-function-docstring
-import os
-import asyncio
-from typing_extensions import Annotated
+# pylint: disable=missing-return-doc, missing-param-doc, missing-function-docstring, duplicate-code
 
-from dotenv import load_dotenv
+import asyncio
+
 import sqlalchemy
 from sqlalchemy import create_engine
 from sqlalchemy.ext.automap import automap_base
 
 import dbally
-from dbally import decorators, SqlAlchemyBaseView
+from dbally import SqlAlchemyBaseView, decorators
 from dbally.audit.event_handlers.cli_event_handler import CLIEventHandler
-from dbally.similarity import SimpleSqlAlchemyFetcher, FaissStore, SimilarityIndex
-from dbally.embeddings.litellm import LiteLLMEmbeddingClient
 from dbally.llms.litellm import LiteLLM
 
-load_dotenv()
 engine = create_engine("sqlite:///examples/recruiting/data/candidates.db")
 
 Base = automap_base()
 Base.prepare(autoload_with=engine)
 
 Candidate = Base.classes.candidates
-
-country_similarity = SimilarityIndex(
-    fetcher=SimpleSqlAlchemyFetcher(
-        engine,
-        table=Candidate,
-        column=Candidate.country,
-    ),
-    store=FaissStore(
-        index_dir="./similarity_indexes",
-        index_name="country_similarity",
-        embedding_client=LiteLLMEmbeddingClient(
-            model="text-embedding-3-small",  # to use openai embedding model
-            api_key=os.environ["OPENAI_API_KEY"],
-        ),
-    ),
-)
 
 
 class CandidateView(SqlAlchemyBaseView):
@@ -69,7 +48,7 @@ class CandidateView(SqlAlchemyBaseView):
         )
 
     @decorators.view_filter()
-    def from_country(self, country: Annotated[str, country_similarity]) -> sqlalchemy.ColumnElement:
+    def from_country(self, country: str) -> sqlalchemy.ColumnElement:
         """
         Filters candidates from a specific country.
         """
@@ -77,14 +56,13 @@ class CandidateView(SqlAlchemyBaseView):
 
 
 async def main():
-    dbally.event_handlers = [CLIEventHandler()]
-    await country_similarity.update()
-
     llm = LiteLLM(model_name="gpt-3.5-turbo")
+    dbally.event_handlers = [CLIEventHandler()]
+
     collection = dbally.create_collection("recruitment", llm)
     collection.add(CandidateView, lambda: CandidateView(engine))
 
-    result = await collection.ask("Find someone from the United States with more than 2 years of experience.")
+    result = await collection.ask("Find me French candidates suitable for a senior data scientist position.")
 
     print(f"The generated SQL query is: {result.context.get('sql')}")
     print()
