+++ examples/semantic_similarity.py
@@ -1,22 +1,44 @@
-# pylint: disable=missing-return-doc, missing-param-doc, missing-function-docstring
-import dbally
+# pylint: disable=missing-return-doc, missing-param-doc, missing-function-docstring, duplicate-code
+
 import asyncio
+import os
 
 import sqlalchemy
+from dotenv import load_dotenv
 from sqlalchemy import create_engine
 from sqlalchemy.ext.automap import automap_base
+from typing_extensions import Annotated
 
-from dbally import decorators, SqlAlchemyBaseView
+import dbally
+from dbally import SqlAlchemyBaseView, decorators
 from dbally.audit.event_handlers.cli_event_handler import CLIEventHandler
+from dbally.embeddings.litellm import LiteLLMEmbeddingClient
 from dbally.llms.litellm import LiteLLM
+from dbally.similarity import FaissStore, SimilarityIndex, SimpleSqlAlchemyFetcher
 
-
+load_dotenv()
 engine = create_engine("sqlite:///examples/recruiting/data/candidates.db")
 
 Base = automap_base()
 Base.prepare(autoload_with=engine)
 
 Candidate = Base.classes.candidates
+
+country_similarity = SimilarityIndex(
+    fetcher=SimpleSqlAlchemyFetcher(
+        engine,
+        table=Candidate,
+        column=Candidate.country,
+    ),
+    store=FaissStore(
+        index_dir="./similarity_indexes",
+        index_name="country_similarity",
+        embedding_client=LiteLLMEmbeddingClient(
+            model="text-embedding-3-small",  # to use openai embedding model
+            api_key=os.environ["OPENAI_API_KEY"],
+        ),
+    ),
+)
 
 
 class CandidateView(SqlAlchemyBaseView):
@@ -48,7 +70,7 @@ class CandidateView(SqlAlchemyBaseView):
         )
 
     @decorators.view_filter()
-    def from_country(self, country: str) -> sqlalchemy.ColumnElement:
+    def from_country(self, country: Annotated[str, country_similarity]) -> sqlalchemy.ColumnElement:
         """
         Filters candidates from a specific country.
         """
@@ -56,13 +78,14 @@ class CandidateView(SqlAlchemyBaseView):
 
 
 async def main():
-    llm = LiteLLM(model_name="gpt-3.5-turbo")
     dbally.event_handlers = [CLIEventHandler()]
+    await country_similarity.update()
 
+    llm = LiteLLM(model_name="gpt-3.5-turbo")
     collection = dbally.create_collection("recruitment", llm)
     collection.add(CandidateView, lambda: CandidateView(engine))
 
-    result = await collection.ask("Find me French candidates suitable for a senior data scientist position.")
+    result = await collection.ask("Find someone from the United States with more than 2 years of experience.")
 
     print(f"The generated SQL query is: {result.context.get('sql')}")
     print()
